wgengine/magicsock: factor out endpoint into its own file
Updates tailscale/corp#13464
